requirements.txt
flask
openai
langchain
python-dotenv
langchain-community
python-docx
webvtt-py